+++ demoapp/src/main/java/com/transferwise/tasks/demoapp/config/TransactionManagerConfigurationHikari.java
@@ -3,21 +3,24 @@ package com.transferwise.tasks.demoapp.config;
 import com.transferwise.common.gaffer.ServiceRegistry;
 import com.transferwise.common.gaffer.ServiceRegistryHolder;
 import com.transferwise.common.gaffer.jdbc.DataSourceImpl;
-import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
+import java.util.Properties;
 import javax.sql.DataSource;
 import javax.transaction.TransactionManager;
 import javax.transaction.UserTransaction;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.core.env.Environment;
 import org.springframework.jdbc.datasource.IsolationLevelDataSourceAdapter;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.transaction.jta.JtaTransactionManager;
 
-//@Configuration
+@Configuration
 @EnableTransactionManagement
+@Slf4j
 public class TransactionManagerConfigurationHikari {
 
   @Autowired
@@ -59,52 +62,59 @@ public class TransactionManagerConfigurationHikari {
   }
 
   private DataSource dataSource() {
-    HikariConfig config = new HikariConfig();
-    config.setJdbcUrl(env.getProperty("spring.datasource.url"));
-    config.setUsername(env.getProperty("spring.datasource.username"));
-    config.setPassword(env.getProperty("spring.datasource.password"));
-    config.setConnectionTimeout(10000);
-    config.setTransactionIsolation("TRANSACTION_READ_COMMITTED");
-    config.setRegisterMbeans(true);
-    config.setMaxLifetime(1800000);
-    config.setMaximumPoolSize(30);
-    config.setMinimumIdle(30);
-    config.setIdleTimeout(60000);
-    config.setValidationTimeout(10000);
+    HikariDataSource hds = new HikariDataSource();
+    hds.setPoolName("demoapp");
+    hds.setJdbcUrl(env.getProperty("spring.datasource.url"));
+    hds.setUsername(env.getProperty("spring.datasource.username"));
+    hds.setPassword(env.getProperty("spring.datasource.password"));
+    hds.setConnectionTimeout(10000);
+    hds.setTransactionIsolation("TRANSACTION_READ_COMMITTED");
+    hds.setRegisterMbeans(true);
+    hds.setMaxLifetime(1800000);
+    hds.setMaximumPoolSize(30);
+    hds.setMinimumIdle(10);
+    hds.setIdleTimeout(60000);
+    hds.setValidationTimeout(10000);
+    hds.setLeakDetectionThreshold(10000);
 
-    if (config.getJdbcUrl().contains("mysql") || config.getJdbcUrl().contains("mariadb")) {
-      config.addDataSourceProperty("cachePrepStmts", "true");
-      config.addDataSourceProperty("prepStmtCacheSize", "250");
-      config.addDataSourceProperty("prepStmtCacheSqlLimit", "4096"); // Long live the Hibernate!
-      config.addDataSourceProperty("useServerPrepStmts", "true");
+    if (hds.getJdbcUrl().contains("mariadb")) {
+      Properties props = hds.getDataSourceProperties();
+      props.put("useServerPrepStmts", "true");
+      props.put("prepStmtCacheSize", "256");
+      props.put("prepStmtCacheSqlLimit", "4096");
+    } else if (hds.getJdbcUrl().contains("mysql")) {
+      hds.addDataSourceProperty("cachePrepStmts", "true");
+      hds.addDataSourceProperty("prepStmtCacheSize", "250");
+      hds.addDataSourceProperty("prepStmtCacheSqlLimit", "4096"); // Long live the Hibernate!
+      hds.addDataSourceProperty("useServerPrepStmts", "true");
 
-      config.addDataSourceProperty("cacheCallableStmts", "true");
-      config.addDataSourceProperty("callableStmtCacheSize", "250");
+      hds.addDataSourceProperty("cacheCallableStmts", "true");
+      hds.addDataSourceProperty("callableStmtCacheSize", "250");
 
-      config.addDataSourceProperty("cacheResultSetMetadata", "true");
-      config.addDataSourceProperty("metadataCacheSize", "250");
+      hds.addDataSourceProperty("cacheResultSetMetadata", "true");
+      hds.addDataSourceProperty("metadataCacheSize", "250");
 
-      config.addDataSourceProperty("useLocalSessionState", "true");
-      config.addDataSourceProperty("useLocalTransactionState", "true"); // Dangerous to have true.
-      config.addDataSourceProperty("cacheServerConfiguration", "true");
-      config.addDataSourceProperty("elideSetAutoCommits", "true");
-      config.addDataSourceProperty("alwaysSendSetIsolation", "false");
+      hds.addDataSourceProperty("useLocalSessionState", "true");
+      hds.addDataSourceProperty("useLocalTransactionState", "true"); // Dangerous to have true.
+      hds.addDataSourceProperty("cacheServerConfiguration", "true");
+      hds.addDataSourceProperty("elideSetAutoCommits", "true");
+      hds.addDataSourceProperty("alwaysSendSetIsolation", "false");
 
-      config.addDataSourceProperty("rewriteBatchedStatements", "true");
-      config.addDataSourceProperty("maintainTimeStats", "false");
-      config.addDataSourceProperty("useJvmCharsetConverters", "true");
-      config.addDataSourceProperty("useSSL", "false");
-      config.addDataSourceProperty("characterEncoding", "UTF-8");
-      config.addDataSourceProperty("useUnicode", "true");
+      hds.addDataSourceProperty("rewriteBatchedStatements", "true");
+      hds.addDataSourceProperty("maintainTimeStats", "false");
+      hds.addDataSourceProperty("useJvmCharsetConverters", "true");
+      hds.addDataSourceProperty("useSSL", "false");
+      hds.addDataSourceProperty("characterEncoding", "UTF-8");
+      hds.addDataSourceProperty("useUnicode", "true");
 
-      config.addDataSourceProperty("dontTrackOpenResources", "true");
-      config.addDataSourceProperty("holdResultsOpenOverStatementClose", "true");
-      config.addDataSourceProperty("enableQueryTimeouts", "false");
+      hds.addDataSourceProperty("dontTrackOpenResources", "true");
+      hds.addDataSourceProperty("holdResultsOpenOverStatementClose", "true");
+      hds.addDataSourceProperty("enableQueryTimeouts", "false");
 
-      config.addDataSourceProperty("connectTimeout", "10000");
-      config.addDataSourceProperty("socketTimeout", "600000");
+      hds.addDataSourceProperty("connectTimeout", "10000");
+      hds.addDataSourceProperty("socketTimeout", "600000");
     }
 
-    return new HikariDataSource(config);
+    return hds;
   }
 }