+++ integration-tests/src/test/java/com/transferwise/tasks/testapp/dao/TaskDaoIntTest.java
@@ -42,6 +42,8 @@ import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
 import lombok.SneakyThrows;
+import org.assertj.core.data.TemporalOffset;
+import org.assertj.core.data.TemporalUnitWithinOffset;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;