+++ tw-tasks-core/src/main/java/com/transferwise/tasks/handler/interfaces/ITaskProcessingPolicy.java
@@ -21,7 +21,18 @@ public interface ITaskProcessingPolicy {
     IGNORE
   }
 
-  default StuckTaskResolutionStrategy getStuckTaskResolutionStrategy(IBaseTask task) {
+  enum StuckDetector {
+    /**
+     * We found `PROCESSING` tasks with the same `processing_client_id` as our `node-id` configuration.
+     */
+    SAME_NODE_STARTUP,
+    /**
+     * The normal cluster wide stuck tasks detector.
+     */
+    CLUSTER_WIDE_STUCK_TASKS_DETECTOR
+  }
+
+  default StuckTaskResolutionStrategy getStuckTaskResolutionStrategy(IBaseTask task, StuckDetector stuckDetector) {
     return StuckTaskResolutionStrategy.MARK_AS_ERROR;
   }